test: 미션별 테스트 메서드 분리
