+++ src/test/java/roomescape/e2e/HtmlTemplateE2ETest.java
@@ -1,0 +1,38 @@
+package roomescape.e2e;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+import static org.hamcrest.CoreMatchers.is;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class HtmlTemplateE2ETest {
+
+    @Test
+    @DisplayName("/admin get 요청 시, 200 OK를 응답해야 한다")
+    void level1() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("/admin/reservation get 요청 시, 200 OK를 응답한 다음, /reservations get을 호출하여 예약 관리 데이터 및 200 OK를 응답해야 한다")
+    void level2() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+}
