+++ src/test/java/roomescape/e2e/LayeredArchitectureE2ETest.java
@@ -1,0 +1,34 @@
+package roomescape.e2e;
+
+import java.lang.reflect.Field;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.test.annotation.DirtiesContext;
+import roomescape.controller.RoomescapeController;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class LayeredArchitectureE2ETest {
+
+    @Autowired
+    private RoomescapeController roomescapeController;
+
+    @Test
+    @DisplayName("컨트롤러는 jdbcTemplate를 의존하지 않아야 한다")
+    void level9() {
+        boolean isJdbcTemplateInjected = false;
+
+        for (Field field : roomescapeController.getClass().getDeclaredFields()) {
+            if (field.getType().equals(JdbcTemplate.class)) {
+                isJdbcTemplateInjected = true;
+                break;
+            }
+        }
+        assertThat(isJdbcTemplateInjected).isFalse();
+    }
+}
