+++ src/test/java/roomescape/e2e/ReservationApiTest.java
@@ -2,9 +2,6 @@ package roomescape.e2e;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import java.lang.reflect.Field;
-import java.sql.Connection;
-import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -15,7 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.RoomescapeController;
 import roomescape.dto.ReservationResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -23,10 +19,8 @@ import static org.hamcrest.CoreMatchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class MissionStepTest {
+public class ReservationApiTest {
 
-    @Autowired
-    private RoomescapeController roomescapeController;
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
@@ -38,30 +32,6 @@ public class MissionStepTest {
         jdbcTemplate.execute("ALTER TABLE reservation_time ALTER COLUMN id RESTART WITH 1");
         jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", "15:40");
         jdbcTemplate.update("INSERT INTO reservation (name, date, time_id) VALUES (?, ?, ?)", "브라운", "2023-08-05", 1L);
-    }
-
-    @Test
-    @DisplayName("/admin get 요청 시, 200 OK를 응답해야 한다")
-    void level1() {
-        RestAssured.given().log().all()
-                .when().get("/admin")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @Test
-    @DisplayName("/admin/reservation get 요청 시, 200 OK를 응답한 다음, /reservations get을 호출하여 예약 관리 데이터 및 200 OK를 응답해야 한다")
-    void level2() {
-        RestAssured.given().log().all()
-                .when().get("/admin/reservation")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
     }
 
     @Test
@@ -97,18 +67,6 @@ public class MissionStepTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(1));
-    }
-
-    @Test
-    @DisplayName("DB 연결을 할 수 있어야 한다")
-    void level4() {
-        try (Connection connection = jdbcTemplate.getDataSource().getConnection()) {
-            assertThat(connection).isNotNull();
-            assertThat(connection.getCatalog()).isEqualTo("DATABASE");
-            assertThat(connection.getMetaData().getTables(null, null, "RESERVATION", null).next()).isTrue();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
     }
 
     @Test
@@ -154,31 +112,6 @@ public class MissionStepTest {
     }
 
     @Test
-    @DisplayName("시간 관리 데이터를 추가/삭제할 수 있다")
-    void level7() {
-        Map<String, String> params = new HashMap<>();
-        params.put("startAt", "10:00");
-
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/times")
-                .then().log().all()
-                .statusCode(200);
-
-        RestAssured.given().log().all()
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(2));
-
-        RestAssured.given().log().all()
-                .when().delete("/times/2")
-                .then().log().all()
-                .statusCode(200);
-    }
-
-    @Test
     @DisplayName("주어진 API 구조를 기반으로 예약 데이터를 추가해야 한다")
     void level8() {
         jdbcTemplate.update("INSERT INTO reservation_time(start_at) values (?)", "11:00");
@@ -199,19 +132,5 @@ public class MissionStepTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(2));
-    }
-
-    @Test
-    @DisplayName("컨트롤러는 jdbcTemplate를 의존하지 않아야 한다")
-    void level9() {
-        boolean isJdbcTemplateInjected = false;
-
-        for (Field field : roomescapeController.getClass().getDeclaredFields()) {
-            if (field.getType().equals(JdbcTemplate.class)) {
-                isJdbcTemplateInjected = true;
-                break;
-            }
-        }
-        assertThat(isJdbcTemplateInjected).isFalse();
     }
 }
